datetime and scheme updates
