+++ Cargo.toml
@@ -10,10 +10,11 @@ edition = "2018"
 futures = "0.3.12"
 confy = "0.4.0"
 serde = "^1.0"
+chrono = "0.4.19"
 
 [dependencies.sqlx]
-# version = "0.4.2"
-path = "../sqlx"
+version = "0.4.2"
+# path = "../sqlx"
 features = [ "mssql", "runtime-async-std-native-tls" ]
 
 [dependencies.async-std]
