+++ src/main.rs
@@ -11,8 +11,9 @@ use futures::TryStreamExt;
 mod schema;
 
 #[async_std::main]
+#[allow(unused_variables)]
 async fn main() -> Result<(), sqlx::Error> {
-
+    // create connection
     let mut conn = MssqlConnectOptions::new()
         .host("HIIWINBL18")
         .username("SNUser")
@@ -20,16 +21,43 @@ async fn main() -> Result<(), sqlx::Error> {
         .connect()
         .await?;
         
-    let query = get_query_from_file("sql/program_status.sql")?;
+    // preload queries
+    let get_status = get_query_from_file("sql/program_status.sql")?;
+    let get_sheet = get_query_from_file("sql/sheet.sql")?;
+    let get_operator = get_query_from_file("sql/operator.sql")?;
 
-    while let Some(input) = get_user_input() {
-    
-        let mut rows = sqlx::query_as::<_, schema::Status>(query.as_str())
+
+    // while let Some(input) = get_user_input() {
+    {
+        let input = "46064";
+        let mut rows = sqlx::query_as::<_, schema::Status>(&get_status)
             .bind(input)
             .fetch(&mut conn);
-    
+
         while let Some(row) = rows.try_next().await? {
-            println!("{:?}", row);
+            // iterates through rows by timestamps in descending order
+            // posting (SN100) and updating (SN102) terminate the loop
+            // deleting (SN101) will contiue the loop because it can be a re-posting or a delete
+            // (re-posting a program sends a SN101 and a SN100, however they do not always go in order)
+
+            // match row.trans_type {
+            //     "SN100" => { // program posted
+            //         println!("Program {} is still active", row.program_name);
+
+            //         break;
+            //     },
+            //     "SN101" => { // program deleted
+            //         println!("Got program deletion.");
+            //     },
+            //     "SN102" => { // program updated
+            //         println!("Updated: {:}", row);
+
+            //         break;
+            //     },
+            //     _ => unreachable!(),
+            // };
+
+            println!("{:}", row);
         }
     }
 
@@ -43,6 +71,7 @@ fn get_query_from_file(file_name: &str) -> Result<String, io::Error> {
     Ok(sql_query)
 }
 
+#[allow(dead_code)]
 fn get_user_input<>() -> Option<String> {
     let mut input = String::new();
